requirements.txt
pymongo[srv]
python-dotenv
h5py
seaborn
matplotlib
pandas
jupyter
python-dateutil

